requirements/production.txt
-r base.txt
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
